requirements.txt
fastapi==0.116.1
pandas==2.3.2
PyJWT==2.10.1
pyodbc==5.2.0
uvicorn==0.22.0
bcrypt==4.2.0
email-validator==2.1.1
python-dotenv==1.0.1
black==24.10.0
ruff==0.6.9
pytest==8.3.3
httpx==0.27.2
pytest-cov==7.0.0
